move install scripts to admin where they belong

+++ misc/admin/oracc-install-funcs.sh
@@ -7,6 +7,9 @@ function dirlinks {
 	mkdir -p $ORACC/$a/$project
 	ln -s $ORACC/$a/$project $ORACC_HOME/$project/02$a	
     done
-    rm -fr $ORACC/$project
-    ln -sf $ORACC_HOME/$project $ORACC/$project
+    subp=`echo $project | tr / -`;
+    if [ "$project" = "$subp" ]; then
+	rm -fr $ORACC/$project
+	ln -sf $ORACC_HOME/$project $ORACC/$project
+    fi
 }
